feat: use x-forwarded-for if config.trust_proxy
If minidyndns is running behind a proxy, then `connection.peeraddr` is
not the true client address. This adds the config `trust_proxy` to
indicate that we can trust `X-Forwarded-For`.

The last address is used because it should be added by the trusted
proxy. I believe this will only work well if there is 1 proxy. If there
are more than one, then we'd have to specify how many, or which proxies
you trust, and pick the address the was provided by the first trusted
proxy. This is out of the scope of this commit though.